requirements.txt
Django==3.2.9
Django==4.2.6
Django==4.2.7
django_paypal==2.0
django_phonenumber_field==7.2.0
Faker==22.0.0
paypal==1.2.5
Pillow==10.1.0
Pillow==9.5.0
Pillow==10.1.0
pycurl==7.45.2
Requests==2.31.0
six==1.16.0
